pyproject.toml
[project]
name = "k-services"
version = "0.1.0"
description = "null"
authors = [
    {name = "matkumatmat",email = "mamattewahyu@gmail.com"}
]
license = {text = "mit"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "gunicorn (>=25.0.1,<26.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "argon2 (>=0.1.10,<0.2.0)",
    "sqlmodel (>=0.0.32,<0.0.33)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "redis (>=7.1.0,<8.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "uuid-utils (>=0.14.0,<0.15.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "maturin (>=1.11.5,<2.0.0)"
]
